Change global xres/yres setting to spin box.

+++ core/mysettings.py
@@ -41,5 +41,5 @@ class MySettings(SettingManager):
         self.addSetting("scale", "integer", "global", 4)
         self.addSetting("rubberWidth", "double", "global", 2)
         self.addSetting("rubberColor", "color", "global", QColor(255, 0, 0))
-        self.addSetting("xres", "float", "global", 30)
-        self.addSetting("yres", "float", "global", 30)
+        self.addSetting("xres", "double", "global", 30)
+        self.addSetting("yres", "double", "global", 30)
